requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0